Refactor IP address calculation to new function
IP address calculation was extracted from the main code and placed into a new function, ipToDecimal, to improve readability and maintainability. This provides a cleaner method to convert an IP address to a decimal, that can be re-used elsewhere if required. Definitions of network, broadcast, subnetMask and ipCount are updated to utilize the new function. In this way, the overall complexity of the main function is reduced.